Revert "Extract top navigation from grid (#23)"
This reverts commit 3048a0662a46bf39e0d0a75e56fdb94d20c1f895.
