+++ src/app/core/grid/grid.component.ts
@@ -1,5 +1,6 @@
 import {
     ChangeDetectionStrategy,
+    ChangeDetectorRef,
     Component,
     EventEmitter,
     HostListener,
@@ -10,9 +11,12 @@ import {
 } from '@angular/core';
 import {MatSlideToggleChange, MatSlideToggleModule} from '@angular/material/slide-toggle';
 import {cloneDeep} from 'lodash';
+import {Subscription, timer} from 'rxjs';
 import {NgFor, NgIf} from '@angular/common';
 import {MatButtonModule} from '@angular/material/button';
 import {GridValueComponent} from '../grid-value/grid-value.component';
+import {MatIconModule} from '@angular/material/icon';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {MatGridListModule} from '@angular/material/grid-list';
 import {SudokuGrid, SudokuRow, SudokuValue} from '../grid-helper/types';
 import {getEmptyRow} from '../grid-helper/empty-row';
@@ -32,16 +36,20 @@ import {NomineeValuesComponent} from '../nominee-values/nominee-values.component
         MatButtonModule,
         MatSlideToggleModule,
         GridValueComponent,
+        MatIconModule,
+        MatTooltipModule,
         MatGridListModule,
         NomineeValuesComponent,
     ],
 })
 class GridComponent implements OnChanges {
     @Input() public originalGrid!: SudokuGrid;
+    @Input() public showTopNavigation: boolean = true;
     @Input() public showFooterNavigation: boolean = true;
-    @Input() public lockValues: boolean = true;
+    @Output() public share: EventEmitter<SudokuGrid> = new EventEmitter<SudokuGrid>();
+    @Output() public create: EventEmitter<void> = new EventEmitter<void>();
+    public lockValues: boolean = true;
     public showNominees: boolean = false;
-    @Output() public onChange: EventEmitter<SudokuGrid> = new EventEmitter<SudokuGrid>();
     protected grid!: SudokuGrid;
     protected solvedGrid: SudokuGrid | null = null;
     protected selectedRowIndex: number | null = null;
@@ -50,6 +58,10 @@ class GridComponent implements OnChanges {
     protected gridNomineeValues: Array<Array<Array<SudokuValue>>> = [];
     protected readonly nomineeValues: SudokuRow = [1, 2, 3, 4, 5, 6, 7, 8, 9];
     @Output() protected finish: EventEmitter<IOnFinishGridEvent> = new EventEmitter<IOnFinishGridEvent>();
+    private time: number = 0;
+    private subscription: Subscription | null = null;
+
+    constructor(private changeDetector: ChangeDetectorRef) {}
 
     public ngOnChanges(changes: SimpleChanges): void {
         if (changes.originalGrid && changes.originalGrid.currentValue !== undefined) {
@@ -159,8 +171,40 @@ class GridComponent implements OnChanges {
         }
     }
 
+    public onChangeLockValues(): void {
+        this.lockValues = !this.lockValues;
+
+        if (this.lockValues) {
+            this.originalGrid = cloneDeep(this.grid);
+            this.initalizeGrid();
+        }
+    }
+
+    public onShareGrid(): void {
+  		this.share.emit(this.originalGrid);
+    }
+
     public trackByIndex(index: number) {
   		return index;
+    }
+
+    public onCreateGrid(): void {
+  		this.create.emit();
+    }
+
+    public timeFormatter(): string {
+  		return timerFormatter(this.time);
+    }
+
+    public clearAllValues(): void {
+        this.grid = [];
+        for (let i: number = 0; i < 9; i++) {
+            this.grid[i] = [];
+
+            for (let k: number = 0; k < 9; k++) {
+                this.grid[i].push(null);
+            }
+        }
     }
 
     private initalizeGrid(): void {
@@ -178,7 +222,15 @@ class GridComponent implements OnChanges {
         this.selectedColIndex = null;
         this.showNominees = false;
         this.isHelpEnabled = false;
+        this.time = 0;
         this.solvedGrid = null;
+
+        this.cancelTimer();
+        this.subscription = timer(0, 1000).subscribe(() => {
+            this.time++;
+
+            this.changeDetector.markForCheck();
+        });
     }
 
     private onValueChange(row: number, col: number, value: SudokuValue): void {
@@ -187,8 +239,6 @@ class GridComponent implements OnChanges {
         this.grid[row][col] = (value > 0 && value <= 9)
             ? value
             : null;
-
-        this.onChange.emit(this.grid);
     }
 
     private toggleNomineeValue(row: number, col: number, value: SudokuValue): void {
@@ -270,13 +320,8 @@ class GridComponent implements OnChanges {
     }
 
     private onFinishGrid(): void {
-        this.finish.emit({
-            grid: this.grid,
-            isGridValid: this.isGridValid(),
-        });
-    }
+        this.cancelTimer();
 
-    private isGridValid(): boolean {
         let isGridValid: boolean = true;
         for (const row of Object.keys(this.grid)) {
             for (const col of Object.keys(this.grid[row])) {
@@ -286,7 +331,17 @@ class GridComponent implements OnChanges {
             }
         }
 
-        return isGridValid;
+        this.finish.emit({
+            grid: this.grid,
+            isGridValid,
+            time: this.time,
+        });
+    }
+
+    private cancelTimer(): void {
+        if (this.subscription) {
+            this.subscription.unsubscribe();
+        }
     }
 
     private isGridFinish(): boolean {
@@ -304,9 +359,27 @@ class GridComponent implements OnChanges {
     }
 }
 
+function timerFormatter(time: number): string {
+    const hours: string = Math
+        .floor(time / 3600)
+        .toString()
+        .padStart(2, '0');
+    const minutes: string = Math
+        .floor(time % 3600 / 60)
+        .toString()
+        .padStart(2, '0');
+    const seconds: string = Math
+        .floor(time % 3600 % 60)
+        .toString()
+        .padStart(2, '0');
+
+    return `${hours}:${minutes}:${seconds}`;
+}
+
 interface IOnFinishGridEvent {
     grid: SudokuGrid;
     isGridValid: boolean;
+    time: number;
 }
 
-export {GridComponent, IOnFinishGridEvent};
+export {GridComponent, IOnFinishGridEvent, timerFormatter};
