+++ src/app/pages/home/home.component.ts
@@ -7,15 +7,11 @@ import {IFinishDialogData, FinishDialogComponent} from '../../core/finish-dialog
 import {
     IOnFinishGridEvent,
     GridComponent,
+    timerFormatter,
 } from '../../core/grid/grid.component';
 import {IShareDialogData, ShareDialogComponent} from '../../core/share-dialog/share-dialog.component';
 import {ClipboardModule} from '@angular/cdk/clipboard';
 import {SudokuGrid, SudokuRow} from '../../core/grid-helper/types';
-import {MatButtonModule} from '@angular/material/button';
-import {MatIconModule} from '@angular/material/icon';
-import {MatTooltipModule} from '@angular/material/tooltip';
-import {NgIf} from '@angular/common';
-import {Subscription, timer} from 'rxjs';
 
 @Component({
     selector: 'home',
@@ -30,18 +26,11 @@ import {Subscription, timer} from 'rxjs';
         FinishDialogComponent,
         CreationDialogComponent,
         ShareDialogComponent,
-        MatButtonModule,
-        MatIconModule,
-        MatTooltipModule,
-        NgIf,
     ],
 })
 class HomeComponent implements OnInit {
-    public lockValues: boolean = true;
-    public sudokuGrid: SudokuGrid = getSudoku('Medium');
+    protected sudokuGrid: SudokuGrid = getSudoku('Medium');
     private readonly lengthOfGridParameter: number = 81;
-    private time: number = 0;
-    private subscription: Subscription | null = null;
 
     constructor(
         private route: ActivatedRoute,
@@ -59,8 +48,6 @@ class HomeComponent implements OnInit {
                 this.changeDetector.markForCheck();
             }
         }
-
-        this.startTime();
     }
 
     public openShareDialog(grid: SudokuGrid): void {
@@ -87,16 +74,12 @@ class HomeComponent implements OnInit {
             })
             .afterClosed()
             .subscribe((difficulty: Difficulty | undefined) => {
-                if (difficulty) {
-                    this.createRandomSudoku(difficulty);
-                    this.startTime();
-                }
+                this.createRandomSudoku(difficulty);
             });
     }
 
     public openFinishDialog(event: IOnFinishGridEvent): void {
-        this.cancelTimer();
-        const time: string = timerFormatter(this.time);
+        const time: string = timerFormatter(event.time);
         const description: string = event.isGridValid
             ? `You solved the puzzle in ${time}`
             : `You did not solve the puzzle in ${time}`;
@@ -129,46 +112,6 @@ class HomeComponent implements OnInit {
         this.sudokuGrid = getSudoku(difficulty);
         this.changeDetector.markForCheck();
     }
-
-    public timeFormatter(): string {
-  	  return timerFormatter(this.time);
-    }
-
-    public clearAllValues(): void {
-        this.sudokuGrid = [];
-        for (let i: number = 0; i < 9; i++) {
-            this.sudokuGrid[i] = [];
-
-            for (let k: number = 0; k < 9; k++) {
-                this.sudokuGrid[i].push(null);
-            }
-        }
-    }
-
-    public onChangeLockValues(): void {
-        this.lockValues = !this.lockValues;
-    }
-
-    public onChangeGrid(grid: SudokuGrid): void {
-        if (!this.lockValues) {
-            this.sudokuGrid = grid;
-        }
-    }
-
-    private cancelTimer(): void {
-        if (this.subscription) {
-            this.subscription.unsubscribe();
-        }
-    }
-
-    private startTime(): void {
-        this.time = 0;
-        this.cancelTimer();
-        this.subscription = timer(0, 1000).subscribe(() => {
-            this.time++;
-            this.changeDetector.markForCheck();
-        });
-    }
 }
 
 function urlParamToGrid(gridString: string): SudokuGrid {
@@ -183,23 +126,6 @@ function urlParamToGrid(gridString: string): SudokuGrid {
 
             return list;
         }, []);
-}
-
-function timerFormatter(time: number): string {
-    const hours: string = Math
-        .floor(time / 3600)
-        .toString()
-        .padStart(2, '0');
-    const minutes: string = Math
-        .floor(time % 3600 / 60)
-        .toString()
-        .padStart(2, '0');
-    const seconds: string = Math
-        .floor(time % 3600 % 60)
-        .toString()
-        .padStart(2, '0');
-
-    return `${hours}:${minutes}:${seconds}`;
 }
 
 export {HomeComponent};
