+++ src/app/pages/page-not-found/page-not-found.component.html
@@ -2,6 +2,7 @@
   <grid
     class="page-not-found__grid"
     [originalGrid]="sudokuGrid"
+    [showTopNavigation]="false"
     [showFooterNavigation]="false"
   ></grid>
